+++ vmf/src/modules/common/output/ComputeStats.hpp
@@ -1,17 +1,8 @@
 /* =============================================================================
  * Vader Modular Fuzzer (VMF)
- * Copyright (c) 2021-2024 The Charles Stark Draper Laboratory, Inc.
+ * Copyright (c) 2021-2025 The Charles Stark Draper Laboratory, Inc.
  * <vmf@draper.com>
- *  
- * Effort sponsored by the U.S. Government under Other Transaction number
- * W9124P-19-9-0001 between AMTC and the Government. The U.S. Government
- * Is authorized to reproduce and distribute reprints for Governmental purposes
- * notwithstanding any copyright notation thereon.
- *  
- * The views and conclusions contained herein are those of the authors and
- * should not be interpreted as necessarily representing the official policies
- * or endorsements, either expressed or implied, of the U.S. Government.
- *  
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 (only) as 
  * published by the Free Software Foundation.
@@ -34,8 +25,11 @@
 namespace vmf
 {
 /**
- * @brief Output module that computes execution statistics and publishes them
- * to the metadata.
+ * @brief OutputModule that computes execution statistics and publishes them
+ * to metadata.
+ * A number of fields are written, for usage by other modules (such as StatsOutput).
+ * @image html CoreModuleDataModel_6.png width=800px
+ * @image latex CoreModuleDataModel_6.png width=6in
  */
 class ComputeStats : public OutputModule {
 public:
@@ -53,7 +47,7 @@ private:
 
     //These are the statistics collected/computed
     ///Total Fuzzer executions
-    unsigned int currentTotal;
+    unsigned long long currentTotal;
     //Total Fuzzer crashes
     unsigned int totalCrashes;
     //Total Fuzzer hangs
@@ -74,7 +68,7 @@ private:
     int outputRate;
     time_t timeLastComputedStats;
     time_t timeLastInterestingTestCaseFound;
-    int prevTestCaseTotal;
+    unsigned long long prevTestCaseTotal;
     double total_time;
 
     //storage handles
